backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
Pillow==10.2.0
numpy==1.26.4
torch==2.2.0
torchvision==0.17.0
ultralytics==8.1.14
opencv-python-headless==4.9.0.80
scikit-image==0.22.0
matplotlib==3.8.3
reportlab==4.1.0
python-jose[cryptography]==3.3.0
pydantic==2.6.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
sqlalchemy==2.0.27
grad-cam==1.4.8
